# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: loginBetaUser
      tags: [auth]
      summary: Sign in to the HeyTelli beta
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthLoginInput"
      responses:
        "200":
          description: Signed-in beta session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Invalid invite code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Return the current signed-in beta user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current beta user
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user:
                    $ref: "#/components/schemas/AuthUser"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: HeyTelli
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: { type: string }
    AuthLoginInput:
      type: object
      required: [email, inviteCode]
      properties:
        email: { type: string, format: email }
        inviteCode: { type: string, minLength: 1 }
        displayName: { type: string, minLength: 1, maxLength: 80 }
    AuthUser:
      type: object
      required: [id, email, displayName, role]
      properties:
        id: { type: integer }
        email: { type: string, format: email }
        displayName: { type: ["string", "null"] }
        role:
          type: string
          enum: [user, admin]
    AuthSession:
      type: object
      required: [token, user]
      properties:
        token: { type: string }
        user:
          $ref: "#/components/schemas/AuthUser"
...