# lib/api-zod/src/generated/types/authLoginInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuthLoginInput {
  email: string;
  /** @minLength 1 */
  inviteCode: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  displayName?: string;
}