# lib/api-zod/src/generated/types/authSession.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AuthUser } from './authUser';

export interface AuthSession {
  token: string;
  user: AuthUser;
}