# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AuthUserRole } from './authUserRole';

export interface AuthUser {
  id: number;
  email: string;
  /** @nullable */
  displayName: string | null;
  role: AuthUserRole;
}